Move esbuild to top level
