+++ package.json
@@ -9,12 +9,16 @@
     "foreach": "yarn workspaces foreach --no-private --parallel --verbose"
   },
   "devDependencies": {
+    "@flex-development/toggle-pkg-type": "^1.0.1",
     "danger": "^11.2.3",
     "duti": "^0.15.2",
+    "esbuild": "^0.17.5",
     "eslint": "^8.33.0",
     "eslint-import-resolver-typescript": "^3.5.3",
     "eslint-plugin-import": "^2.27.5",
+    "glob": "^8.1.0",
     "jest": "^29.4.1",
+    "minimist": "^1.2.7",
     "prettier": "^2.8.3"
   },
   "packageManager": "yarn@3.3.1"
