+++ packages/client/package.json
@@ -27,7 +27,7 @@
     }
   },
   "scripts": {
-    "prepack": "node scripts/build.js"
+    "prepack": "node ../../scripts/esbuild.js --platform=browser"
   },
   "repository": {
     "type": "git",
@@ -44,9 +44,6 @@
     "lodash": "^4.17.15"
   },
   "devDependencies": {
-    "@flex-development/toggle-pkg-type": "^1.0.1",
-    "esbuild": "^0.16.17",
-    "glob": "^8.0.3",
     "mobx": "^4.3.1"
   },
   "packageManager": "yarn@3.3.1"
