+++ packages/provider-elasticsearch/package.json
@@ -27,7 +27,7 @@
     }
   },
   "scripts": {
-    "prepack": "node scripts/build.js"
+    "prepack": "node ../../scripts/esbuild.js --platform=node"
   },
   "repository": {
     "type": "git",
@@ -55,11 +55,8 @@
   },
   "devDependencies": {
     "@elastic/elasticsearch": "^7.11.0",
-    "@flex-development/toggle-pkg-type": "^1.0.1",
     "agentkeepalive": "^4.1.4",
     "contexture": "^0.12.11",
-    "esbuild": "^0.16.17",
-    "glob": "^8.0.3",
     "json-stable-stringify": "^1.0.1"
   },
   "packageManager": "yarn@3.2.0"
