+++ packages/provider-mongo/package.json
@@ -27,7 +27,7 @@
     }
   },
   "scripts": {
-    "prepack": "node scripts/build.js"
+    "prepack": "node ../../scripts/esbuild.js --platform=node"
   },
   "repository": {
     "type": "git",
@@ -54,10 +54,7 @@
     "moment-timezone": "^0.5.28"
   },
   "devDependencies": {
-    "@flex-development/toggle-pkg-type": "^1.0.1",
     "contexture": "^0.0.4",
-    "esbuild": "^0.16.17",
-    "glob": "^8.0.3",
     "mingo": "^4.1.2",
     "mongodb": "^3.5.6",
     "mongodb-memory-server": "^6.5.2"
