--- packages/react/.vscode/launch.json
@@ -1,17 +1,0 @@
-{
-  // Use IntelliSense to learn about possible attributes.
-  // Hover to view descriptions of existing attributes.
-  // For more information, visit: https://go.microsoft.com/fwlink/?linkid=830387
-  "version": "0.2.0",
-  "configurations": [
-    {
-      // You need to install the Debugger for Chrome extension for this to work.
-      // To start debuggin make sure to run `npm start` in the console first.
-      "type": "chrome",
-      "request": "launch",
-      "name": "Test Storybooks",
-      "url": "http://localhost:3000",
-      "webRoot": "${workspaceFolder}/src"
-    }
-  ]
-}
