+++ packages/react/package.json
@@ -27,7 +27,7 @@
     }
   },
   "scripts": {
-    "prepack": "node scripts/build.js",
+    "prepack": "node ../../scripts/esbuild.js --platform=browser --loader:.js=jsx",
     "start": "npm run storybook",
     "storybook": "start-storybook -p 3001",
     "storybook:build": "build-storybook",
@@ -57,7 +57,6 @@
   },
   "devDependencies": {
     "@date-io/moment": "^1.3.9",
-    "@flex-development/toggle-pkg-type": "^1.0.1",
     "@material-ui/core": "^4.3.3",
     "@material-ui/pickers": "^3.2.3",
     "@storybook/addon-actions": "^6.3.7",
@@ -70,9 +69,7 @@
     "contexture-elasticsearch": "^1.13.0",
     "elasticsearch-browser": "^14.2.2",
     "emoji-datasource": "^5.0.1",
-    "esbuild": "^0.16.17",
     "eslint-plugin-react": "^7.32.0",
-    "glob": "^8.0.3",
     "material-ui-chip-input": "^1.0.0",
     "mobx": "^4.3.1",
     "mobx-react": "^6.3.0",
