+++ packages/server/package.json
@@ -27,7 +27,7 @@
     }
   },
   "scripts": {
-    "prepack": "node scripts/build.js"
+    "prepack": "node ../../scripts/esbuild.js --platform=node"
   },
   "repository": {
     "type": "git",
@@ -52,9 +52,6 @@
     "moment-timezone": "^0.5.28"
   },
   "devDependencies": {
-    "@flex-development/toggle-pkg-type": "^1.0.1",
-    "esbuild": "^0.15.12",
-    "glob": "^8.0.3",
     "mockdate": "^3.0.5"
   },
   "packageManager": "yarn@3.2.4"
